2011-01-30	Ralf Corsépius <ralf.corsepius@rtems.org>
	* mongoosev/duart/mg5uart.c: Fix type mismatch.

+++ c/src/lib/libcpu/mips/mongoosev/duart/mg5uart.c
@@ -633,7 +633,7 @@ MG5UART_STATIC void mg5uart_initialize_interrupts(int minor)
 MG5UART_STATIC int mg5uart_write_support_int(
   int         minor,
   const char *buf,
-  int         len
+  size_t      len
 )
 {
   uint32_t        Irql;
